fix: US-005 - Split langchain vectorstore.py into focused modules
Extract SearchOpsMixin (15 search methods) into search_ops.py and
GraphOpsMixin (query, explain, match_query) into graph_ops.py.
VelesDBVectorStore inherits both mixins via MRO.
vectorstore.py: 495 NLOC, search_ops.py: 451 NLOC, graph_ops.py: 70 NLOC.
No assert statements. 80 tests pass (1 pre-existing version test failure).
